refactor: extract now-playing message/journal service helpers
Move the self-contained async helpers deleteLatestJournalMessageInChannel,
trackJournalReply, and deleteEligibleNowPlayingMessageInCurrentChannel into
now-playing/nowPlayingMessageService.ts. These have no instance state and no
cross-handler this calls; call sites updated to imports.

Refs #894
